requirements.txt
streamlit==1.14.0
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
torch